+++ src/views/page-header.js
@@ -4,13 +4,20 @@ import usage from './page-header.usage.js';
 
 export default function(discovery) {
     discovery.view.define('page-header', function render(el, config, data, context) {
-        const { prelude, content } = config;
+        const { prelude, content, onInit, onChange } = config;
         const preludeEl = el.appendChild(createElement('div', 'view-page-header__prelude'));
         const contentEl = el.appendChild(createElement('div', 'view-page-header__content'));
+        const mixinHandlers = (config) =>
+            typeof onInit !== 'function' && typeof onChange !== 'function'
+                ? config // left as is since nothing to mix in
+                : this.composeConfig(config, {
+                    onInit,
+                    onChange
+                });
 
         return Promise.all([
-            discovery.view.render(preludeEl, prelude || [], data, context),
-            discovery.view.render(contentEl, content || 'text', data, context)
+            discovery.view.render(preludeEl, mixinHandlers(prelude || []), data, context),
+            discovery.view.render(contentEl, mixinHandlers(content || 'text'), data, context)
         ]);
     }, { usage });
 }